Refactor updater script to handle missing music directory

+++ fnancz2-updater.sh
@@ -34,7 +34,8 @@ if [[ "$version" == "0.9.11" ]]; then
    read -n 1 -s -p "Press any key to continue..."
 elif [[ "$version" == "0.9.10" ]]; then
    echo -e "\033[34mUpdating the game files...\033[0m"
-     if wget https://github.com/AmrThePigeon/FNANCZAE2_Script_Builder/releases/download/Updates/0.9.10-to-0.9.11.xdelta; then 
+    if [[ ! -d "assets/music_n_sfx" ]]; then
+      if wget https://github.com/AmrThePigeon/FNANCZAE2_Script_Builder/releases/download/Updates/0.9.10-to-0.9.11.xdelta; then 
         if mv "assets/game.unx" "assets/data.win"; then
            if xdelta3 -d -s "assets/data.win" "0.9.10-to-0.9.11.xdelta" "assets/game.unx"; then
               rm "0.9.10-to-0.9.11.xdelta" 2>/dev/null
@@ -45,7 +46,20 @@ elif [[ "$version" == "0.9.10" ]]; then
               exit 1
            fi
         fi
-     fi 
+     fi
+  else
+      if wget https://github.com/AmrThePigeon/FNANCZAE2_Script_Builder/releases/download/Updates/0.9.10-to-0.9.11(mod).xdelta; then 
+        if mv "assets/game.unx" "assets/data.win"; then
+           if xdelta3 -d -s "assets/data.win" "0.9.10-to-0.9.11(mod).xdelta" "assets/game.unx"; then
+              rm "0.9.10-to-0.9.11(mod).xdelta" 2>/dev/null
+              rm "assets/data.win" 2>/dev/null
+              echo -e "\033[34mUpdate files have been applied. The game was updated to latest version successfully (v0.9.11)\033[0m"
+              echo -e "0.9.11" > "assets/versioninfo.txt"
+              read -n 1 -s -p "Press any key to continue..."
+              exit 1       
+           fi
+        fi
+     fi
 #elif [[ "$version" == "0.9.9" ]]; then (Version 0.9.9 no longer obtainable)
 #   echo -e "\033[34mUpdating the game files...\033[0m"
 #     if #Download of v0.9.10 xdelta command here && #Download command of v9.11 xdelta here; then
